qMax is picky about its argument types.
svn path=/trunk/KDE/kdelibs/; revision=1010050

+++ kioslave/http/http_cache_cleaner.cpp
@@ -86,8 +86,8 @@ struct CacheFileInfo {
 
 bool CacheFileInfo::operator<(const CacheFileInfo &other) const
 {
-    const int thisUseful = useCount / qMax(g_currentDate - lastUsedDate, 1);
-    const int otherUseful = other.useCount / qMax(g_currentDate - other.lastUsedDate, 1);
+    const int thisUseful = useCount / qMax(g_currentDate - lastUsedDate, qint64(1));
+    const int otherUseful = other.useCount / qMax(g_currentDate - other.lastUsedDate, qint64(1));
     return thisUseful < otherUseful;
 }
